put run gcovr step in a independent function
